backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
xmltodict==0.13.0
pandas==2.1.4
openpyxl==3.1.2
